Complete comprehensive test suite expansion with modular external tests
This commit completes the comprehensive test suite expansion following the user's
request to "read the blueprints in spec, and look at the test/ and src/test/ to
see how we can expand our unit tests for all functions as well as the overall
shell based comprehensive test". The expansion includes:

## Unit Test Improvements:
- Fixed compilation errors in existing test files
- Updated test files to use correct struct members (opts.exec_argv, g_state.verbose, etc.)
- Fixed format constant references (FMT_HUMAN instead of FORMAT_HUMAN)
- Cleaned up unit_tests.c to be a proper test coordinator
- Fixed type casting warnings in checksum tests

## Comprehensive External Test Suite:
Created 12 focused external test files organized by functionality:
- test_help_version.sh: Help and version command tests
- test_list.sh: Lock listing functionality tests
- test_check.sh: Lock checking functionality tests
- test_mutex.sh: Basic mutex functionality tests
- test_semaphore.sh: Semaphore functionality tests
- test_timeout.sh: Timeout functionality tests
- test_done.sh: Done signal functionality tests
- test_exec.sh: Exec mode functionality tests
- test_onepercpu.sh: One-per-CPU functionality tests
- test_excludecpus.sh: CPU exclusion functionality tests
- test_environment.sh: Environment variable tests
- test_scenarios.sh: Real-world scenario tests

## Test Framework:
- test_framework.sh: Common framework with helper functions
- run_all_tests.sh: Main test runner with individual test execution
- Color-coded output and comprehensive test reporting
- Test isolation with proper cleanup
- Modular design allowing focused testing

## Test Coverage:
- All command-line options and variations
- Real-world usage scenarios from README examples
- Error conditions and edge cases
- Performance and stress testing scenarios
- Multi-process concurrency testing
- Signal handling and cleanup verification

This addresses the user's feedback to organize tests "not only by parameter,
but also scenario, the way the examples in the README are built" and creates
a comprehensive test suite that validates all waitlock functionality.

+++ src/test/test_checksum.c
@@ -117,7 +117,7 @@ int test_crc32_large_data(void) {
     int i;
     
     /* Fill with pattern */
-    for (i = 0; i < sizeof(large_data); i++) {
+    for (i = 0; i < (int)sizeof(large_data); i++) {
         large_data[i] = (char)(i % 256);
     }
